experiment/portfolio_rebalancer/requirements.txt
flask==2.3.3
flask-cors==4.0.0
web3==6.11.1
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2 